more work on party affiliation facet and mapping
there's a lot of bad data in this field in the voting record
so if we find a null value or a unmappable id we'll just skip
indexing it as a facet since there are a lot of them and it'll make a
mess.  Erik will work on data clean up and this will progressively
improve.